app/requirements.txt
flask==2.0.1
werkzeug==2.0.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-cosmos==4.3.0
msrest==0.7.1
python-dotenv==0.19.1
pyodbc==4.0.34
SQLAlchemy==1.4.46